canvas: move Refresh() into canvas.go
