+++ canvas/base.go
@@ -98,15 +98,3 @@ func (o *baseObject) Visible() bool {
 
 	return !o.Hidden
 }
-
-// Refresh instructs the containing canvas to refresh the specified obj.
-func Refresh(obj fyne.CanvasObject) {
-	if fyne.CurrentApp() == nil || fyne.CurrentApp().Driver() == nil {
-		return
-	}
-
-	c := fyne.CurrentApp().Driver().CanvasForObject(obj)
-	if c != nil {
-		c.Refresh(obj)
-	}
-}
